backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
bcrypt==5.0.0
cffi==1.17.1
click==8.1.8
cryptography==46.0.3
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.109.0
greenlet==3.1.1
h11==0.16.0
httptools==0.6.4
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
Mako==1.3.10
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.4.8
pycparser==2.23
pydantic==2.10.6
pydantic-settings==2.1.0
pydantic_core==2.27.2
python-dotenv==1.0.0
python-jose==3.4.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.13.2
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2
